+++ frontend/src/routes/settings/index.tsx
@@ -1,4 +1,3 @@
-import { SearchBox } from "@/components/SearchBox";
 import { requireAuth } from "@/lib/requireAuth";
 import { createFileRoute } from "@tanstack/react-router";
 
@@ -13,7 +12,6 @@ function SettingsIndex() {
   return (
     <div className="flex items-center justify-between mb-4">
       <h1 className="text-2xl font-bold">Settings</h1>
-      <SearchBox />
     </div>
   );
 }
